Type System started
In order to contain the mess of reflection magic spilling all over the code
base, GoLX now has the beginings of a type system independent of Go's. It
is designed to be easily interoperable with Go types and at this stage all
GoLX types effectively have to be Go types as well.